Destructes more small parts of App Router
