+++ src/js/navigation/Routes.js
@@ -1,0 +1,23 @@
+import React, { Component } from 'react';
+
+class Routes extends Component {
+    
+    render() {
+        
+        var routes = this.props.pageRefs.map((ref, index) => {
+            
+            // Creates Route elements for each page.
+            if (index === 0) {
+                return (<Route key={ref.name + "-route"} path="/" exact component={ref.page} />);
+            } else {
+                return (<Route key={ref.name + "-route"} path={ref.url} component={ref.page}/>);
+            }
+        });
+        
+        return(
+            {routes}
+        );
+    }
+}
+
+export default Routes;
